API | Customization des réponses API

+++ backend/api/viewsets.py
@@ -17,8 +17,8 @@ def Connexion(request):
 	""" Cette fonction prend en charge de la connexion des utilisateurs.
 	Methode autorisée: POST,
 	JSON à soumettre: {
-		"login": "...",
-		"password": "..."
+		"login": "...", // Type String/Str
+		"password": "..." // Type String/Str
 	}
 	"""
 	# Vérification de la validité des données collectées
@@ -70,79 +70,100 @@ def isDisconnected(request):
 	pour le déconnecter
 	Méthode autorisée: POST,
 	JSON à soumettre: {
-		"token": "...",
-		"signature": "..."
+		"token": "...", // Type String/Str
+		"signature": "..." // Type String/Str
 	}
 	Ces informations sont fournis lors de la connexion
 	"""
 	try:
-		# Récupération de l'utilisateur
-		user = User.objects.get(pseudo=jwt.decode(request.data['token'], os.environ.get('JWT_SECRET'), algorithms="HS256")['sub'])
-		# Boucle de vérification ( à voir si c'est performant ou pas)
+	# Boucle de vérification ( à voir si c'est performant ou pas)
+		id_user = User.objects.get(pseudo=jwt.decode(request.data['token'], os.environ.get('JWT_SECRET'), algorithms="HS256")['sub']).id
 		while True:
 			# En attente de changement du token (c'est à dire, une nouvelle connexion)
+			# Récupération de l'utilisateur
+			user = User.objects.get(id=id_user)
+			if not user.is_active:
+				print("You need to validate the code")
+				return Response({'status': True, 'motif': "Validate Code"})
+			print('User is already active')
 			if request.data['signature'] != Token.objects.filter(user=user)[0].key:
-				break
-			time.sleep(10)
-		return Response({'status': True})
+				print("You need to reconnect")
+				return Response({'status': True, 'motif': "New Connexion"})
+			time.sleep(10)	
 	except:
-		return Response({'status': 'FAILED'})
+		return Response({'status': 'FAILED', 'message':'Token Invalide'})
 
 @api_view(['POST'])
 def CreateAccountViewset(request):
-	""" Cette fonction, permet de creer un compte utilisateur 
-	Méthode autorisée: POST,
-	JSON à soumettre: {"pseudo": "...", "password": "...", "email":"..."} ou
-					  {"pseudo": "...", "password": "...", "phone":"..."}
-	"""
-	data = request.data
-	keys = list(data.keys())
-	# Vérification de la validité des données collectées
-	if (len(keys) != 3) or ('pseudo' not in keys) or ('password' not in keys) or (('phone' not in keys) and ('email' not in keys)):
-		return Response({'status': 'FAILED'})
+	try:
+		""" Cette fonction, permet de creer un compte utilisateur 
+		Méthode autorisée: POST,
+		JSON à soumettre: 
+		{
+		"pseudo": "...", // Type String/Str
+		"password": "...", // Type String/Str
+		"email":"..." // Type String/Str
+		} ou
+		{
+		"pseudo": "...", // Type String/Str
+		"password": "...", // Type String/Str
+		"phone":"..." // Type String/Str
+		}
+		"""
+		data = request.data
+		keys = list(data.keys())
+		# Vérification de la validité des données collectées
+		if (len(keys) != 3) or ('pseudo' not in keys) or ('password' not in keys) or (('phone' not in keys) and ('email' not in keys)):
+			return Response({'status': 'FAILED', 'message':'JSON invalide'})
 
-	if 'email' in keys:
-		# L'utilisateur a donné son email
-		email = data['email']
-		emailRegex = "([A-Za-z0-9]+[._-]?)*[A-Za-z0-9]+@([A-Za-z0-9]+[._-]?)*[A-Za-z0-9]+\\.([A-Z|a-z]{2,})"
-		if(not re.match(emailRegex, email)): # Regex Email Vérification
-			return Response({'status': 'Email Invalide'})
-		userExist, response = verifyUser(email) # Vérification d'un potentiel utilisateur avec cet email
-	else:
-		#L'utilisateur a donné son phone
-		phone =  data['phone']
-		phoneRegex = "^(77|78|75|70|76)[0-9]{7}$" # Regex Phone Verification
-		if(not re.match(phoneRegex, phone)):
-			return Response({'status': 'Phone Invalide'})
-		userExist, response = verifyUser(data['phone']) # Vérification d'un potentiel utilisateur avec ce mail
-	if userExist:
-		return response
-	serializer = CreateAccountSerializer(data=data) # Sérialisation
-	if not serializer.is_valid():
-		return Response({'status': 'FAILED'})
-	serializer.save()
-	# Envoie du code de Vérification et Création du Payload JWT
-	if 'email' in keys:
-		code = sendVerificationCodeByMail(email)
-		payload = createValidationTokenPayload(code, email, "email")
-	else:
-		code = sendVerificationCodeBySms(phone)
-		payload = createValidationTokenPayload(code, phone, "phone")
-	return Response({
-		'status': "SUCCESSFUL",
-		'token': createToken(payload)
-	})
+		if 'email' in keys:
+			# L'utilisateur a donné son email
+			email = data['email']
+			emailRegex = "([A-Za-z0-9]+[._-]?)*[A-Za-z0-9]+@([A-Za-z0-9]+[._-]?)*[A-Za-z0-9]+\\.([A-Z|a-z]{2,})"
+			if(not re.match(emailRegex, email)): # Regex Email Vérification
+				return Response({'status': 'Email Invalide'})
+			userExist, response = verifyUser(email) # Vérification d'un potentiel utilisateur avec cet email
+		else:
+			#L'utilisateur a donné son phone
+			phone =  data['phone']
+			phoneRegex = "^(77|78|75|70|76)[0-9]{7}$" # Regex Phone Verification
+			if(not re.match(phoneRegex, phone)):
+				return Response({'status': 'Phone Invalide'})
+			userExist, response = verifyUser(data['phone']) # Vérification d'un potentiel utilisateur avec ce mail
+		if userExist:
+			return response
+		serializer = CreateAccountSerializer(data=data) # Sérialisation
+		if not serializer.is_valid():
+			return Response({'status': 'FAILED', 'message': 'Types des données du JSON invalides!'})
+		serializer.save()
+		# Envoie du code de Vérification et Création du Payload JWT
+		if 'email' in keys:
+			code = sendVerificationCodeByMail(email)
+			payload = createValidationTokenPayload(code, email, "email")
+		else:
+			code = sendVerificationCodeBySms(phone)
+			payload = createValidationTokenPayload(code, phone, "phone")
+		return Response({
+			'status': "SUCCESSFUL",
+			'token': createToken(payload)
+		})
+	except:
+		Response({'status': 'FAILED', 'message': "Vérifier votre connexion, Si l'erreur persiste, contactez moi!"})
 
 @api_view(['POST'])
 def ValidateCodeViewset(request):
 	""" Cette fonction, permet de valider le code de l'utilisateur 
 	Méthode autorisée: POST,
-	JSON à soumettre: {"code": "...", "token": "..."}
+	JSON à soumettre: 
+	{
+	"code": ..., // Type Number/int
+	"token": "..." // Type String/str
+	}
 	"""
 	try:
 		isValid, userId = verifyCodeValidation(request.data['code'], request.data['token'])
 		if not isValid:
-			return Response({'status': "FAILED", 'message': 'Invalide Code'})
+			return Response({'status': "FAILED", 'message': 'Token ou Code Invalide'})
 		if User.objects.filter(email=userId):
 			user = User.objects.filter(email=userId)[0]
 		if User.objects.filter(phone=userId):
@@ -151,7 +172,7 @@ def ValidateCodeViewset(request):
 		user.save()
 		return Response({'status': "SUCCESSFUL"})
 	except:
-		return Response({'status': "FAILED"})
+		return Response({'status': "FAILED", 'message':'Erreur non identifié !'})
 	
 
 class PaymentMethodViewset(APIView):
@@ -165,17 +186,22 @@ class PaymentMethodViewset(APIView):
 	def post(self, request):
 		""" Permet d'ajouter une méthode de paiement
 		Méthode: POST,
-		JSON: {'token': '...', 'signature': '...', 'name':'...', 'phone': '...'}
+		JSON: {
+		'token': '...', // Type String/str 
+		'signature': '...',  // Type string/str
+		'name':'...', // Type String/str | Valeurs (WAVE, OM, FREE)
+		'phone': ... // Type Number/Int
+		}
 		"""
 		try:
 			# Validité des éléments
 			if len(request.data) != 4:
-				return Response({'status': "FAILED"})
+				return Response({'status': "FAILED", 'message':"JSON invalide"})
 			token = request.data['token']
 
 			# Vérifie si l'utilisateur est connecté
 			if not isAuthenticated(token, request.data['signature']):
-				return Response({"status": "FAILED"})
+				return Response({"status": "FAILED", 'message': "Vous devez vous connecter"})
 			
 			# Verification existence PM
 			if self.verifyExistingPm(request.data['name'], request.data['phone']):
@@ -187,30 +213,37 @@ class PaymentMethodViewset(APIView):
 			serializer = PaymentMethodSerializer(data=data)
 			if serializer.is_valid():
 				serializer.save()
-				return Response({'status': "SUCCESSFUL"})
+				data = serializer.data
+				data['id'] = PaymentMethod.objects.get(name=data['name'], phone=data['phone']).id
+				return Response({'status': "SUCCESSFUL", "paymentMethod": PaymentMethodForConnSerializer(data).data})
+			return Response({'status': "FAILED", "message":"Types des données du JSON invalides!"})
 		except:
-			return Response({'status': "FAILED"})
-		return Response({'status': "FAILED"})
+			return Response({'status': "FAILED", "message":"Token ou Signature Invalide"})
 		
 
 	def delete(self, request):
 		""" Permet de supprimer une méthode de paiement
 		Méthode: POST,
-		JSON: {'token': '...', 'signature': '...', 'id':'...'}
+		JSON: 
+		{
+		'token': '...', // Type String/str
+		'signature': '...', // Type String/str
+		'id':... // Type Number/int
+		}
 		"""
 		try:
 			# Validité des éléments
 			if len(request.data) != 3:
-				return Response({'status': "FAILED"})
+				return Response({'status': "FAILED", "message":"JSON invalide"})
 			token = request.data['token']
 			# Vérifie si l'utilisateur est connecté
 			if not isAuthenticated(token, request.data['signature']):
-				return Response({"status": "FAILED"})
+				return Response({"status": "FAILED", "message":"Vous devez vous connecter"})
 			pm = PaymentMethod.objects.get(id=int(request.data['id']))
 			pm.delete()
 			return Response({'status': "SUCCESSFUL"})
 		except Exception as e:
-			return Response({'status': "FAILED"})
+			return Response({'status': "FAILED", "message":"Token ou Signature Invalide"})
 
 @api_view(['PATCH'])
 def SetUserViewset(request):
@@ -218,13 +251,13 @@ def SetUserViewset(request):
 	Permet de modifier les informations de l'utilisateurs
 	Méthode: PATCH,
 	JSON: {
-		"token": "...", 
-		"signature":"...", 
-		"pseudo":"...", 
-		"email":"...", 
-		"password": "...", 
-		"newPassword": "...", 
-		"currency": "..."
+		"token": "...", // Type String/str
+		"signature":"...", // Type String/str
+		"pseudo":"...", // Type String/str
+		"email":"...", // Type String/str
+		"password": "...", // Type String/str
+		"newPassword": "...", // Type String/str
+		"currency": "..." // Type String/str
 	}
 	"""
 	try:
@@ -233,13 +266,13 @@ def SetUserViewset(request):
 
 		# Validité des éléments
 		if len(data) > 7:
-			return Response({'status': "FAILED"})
+			return Response({'status': "FAILED", "message":"JSON invalide"})
 
 		token = data['token'] # Récupération du Token
 
 		# Vérifie si l'utilisateur est connecté
 		if not isAuthenticated(token, data['signature']):
-			return Response({"status": "FAILED"})
+			return Response({"status": "FAILED", "message": "Vous devez vous connecter"})
 
 		user = User.objects.get(pseudo=jwt.decode(token, os.environ.get('JWT_SECRET'), algorithms="HS256")['sub'])
 		
@@ -266,7 +299,7 @@ def SetUserViewset(request):
 				return Response({'status': 'Phone Invalide'})
 
 		if (('password' in keys) and ('newPassword' not in keys)) or (('newPassword' in keys) and ('password' not in keys)):
-			return Response({"status":"FAILED"})
+			return Response({"status":"FAILED", "message":" Les deux champs password et newPassword sont obligatoires"})
 
 		if ('password' in keys) and ('newPassword' in keys):
 			if user.check_password(data['password']):
@@ -283,7 +316,7 @@ def SetUserViewset(request):
 		serializer = SetAccountSerializer(user, data=data, partial=True)
 		if not serializer.is_valid():
 			#return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-			return Response({"status":"FAILED"})
+			return Response({"status":"FAILED", "message": "Les Types des données du JSON sont invalides"})
 
 		# Préparation des payload et envoie du code
 		if (('email' in keys) and (user.email != data['email'])) or (('phone' in keys) and (user.phone != str(data['phone']))):
@@ -294,19 +327,27 @@ def SetUserViewset(request):
 				code = sendVerificationCodeBySms(date['phone'])
 				payload = createValidationTokenPayload(code, date['phone'], "phone")
 			serializer.save()
-			user.disconnect()
-			Token.objects.filter(user=user)[0].delete()
-			Token.objects.create(user=user)
+			user.is_active = False
+			user.save()
 			return Response({
 				'status': "SUCCESSFUL",
 				'token': createToken(payload)
 			})
+		if (('pseudo' in keys) and (user.pseudo != data['pseudo'])) or ('password' in keys):
+			serializer.save()
+			Token.objects.filter(user=user)[0].delete()
+			Token.objects.create(user=user)
+			return Response({
+				'status': "SUCCESSFUL"
+			})
 		serializer.save()
-		return Response({
-			'status': "SUCCESSFUL"
-		})
+		if ('currency' in keys):
+			return Response({
+				'status': "SUCCESSFUL",
+				'currency': user.currency
+			})
 	except:
-		return Response({"status":"FAILED"})
+		return Response({"status":"FAILED", "message":"Token ou Signature Invalide ou Vous n'etes pas connecté à internet"})
 
 
 
@@ -318,20 +359,20 @@ class AdViewset(APIView):
 		Méthode: PATCH,
 		JSON: {
 			// Obligatoire
-			"token": "...", 
-			"signature":"...", 
-			"sens":"...", 
-			"quantityType":"...", 
-			"amountType": "...", 
+			"token": "...", // Type String/str 
+			"signature":"...", // Type String/str
+			"sens":"...", // Type String/str | values(A ou V)
+			"quantityType":"...", // Type String/str | values(F ou R)
+			"amountType": "...", // Type String/str | values(F ou R)
 			"marge": "...", //Number, Int
 			"provider": "..."
 			// Optionnel selon le Type
-			'quantityFixe': '...',
-			'quantityMin': '...',
-			'quantityMax': '...',
-			'amountFixe': '...',
-			'amountMin': '...',
-			'amountMax': '...'
+			'quantityFixe': '...', // Type String/str
+			'quantityMin': '...', // Type String/str
+			'quantityMax': '...', // Type String/str
+			'amountFixe': '...', // Type String/str
+			'amountMin': '...', // Type String/str
+			'amountMax': '...' // Type String/str
 		}
 		"""
 		try:
@@ -340,7 +381,7 @@ class AdViewset(APIView):
 			
 			# Vérifie si l'utilisateur est connecté
 			if not isAuthenticated(token, data['signature']):
-				return Response({"status": "FAILED"})
+				return Response({"status": "FAILED", "message": "Vous devez vous connecter"})
 			
 			# Nettoyage de data
 			data.pop('token')
@@ -352,24 +393,21 @@ class AdViewset(APIView):
 			# Donnée en plus
 			for key in keys:
 				if key not in fields:
-					pprint(key)
-					pprint('Not in keys')
-					return Response({"status": "FAILED"})
+					return Response({"status": "FAILED", "message":"JSON invalide"})
 
 			# Vérification de la conformité des données
 			if data['quantityType'] != data['amountType']:
-				return Response({"status": "FAILED"})
-			pprint('not same')
-			if (data['quantityType'] == 'F') and (('quantityFixe' not in keys) or ('quantityFixe' not in keys)):
-				return Response({"status": "FAILED"})
+				return Response({"status": "FAILED", "message":"quantityType et amountType ne peuvent pas être différents"})
+			if (data['quantityType'] == 'F') and (('quantityFixe' not in keys) or ('amountFixe' not in keys)):
+				return Response({"status": "FAILED", "message":"Quand le type est fixe, les champs quantityFixe et amountFixe deviennent Obligatoires"})
 			if (data['quantityType'] == 'F') and (('quantityMin' in keys) or ('quantityMax' in keys) or ('amountMin' in keys) or ('amountMax' in keys)):
-				return Response({"status": "FAILED"})
+				return Response({"status": "FAILED", "message":"Quand le type est fixe, les champs (quantityMin, quantityMax, amountMin, amountMax) ne doivent pas figurer dans le JSON"})
 			if (data['quantityType'] == 'R') and (('quantityMin' not in keys) or ('quantityMax' not in keys) or ('amountMin' not in keys) or ('amountMax' not in keys)):
-				return Response({"status": 'FAILED'})
+				return Response({"status": 'FAILED', "message":"Quand le type est range, les champs (quantityMin, quantityMax, amountMin, amountMax) deviennent Obligatoires"})
 			if (data['quantityType'] == 'R') and (('quantityFixe' in keys) or ('quantityFixe' in keys)):
-				return Response({"status": 'FAILED'})
+				return Response({"status": 'FAILED', "message":"Quand le type est range, les champs quantityFixe et amountFixe ne doivent pas figurer dans le JSON"})
 			if (data['quantityType'] == 'R') and ((data['quantityMin'] >= data['quantityMax']) or (data['amountMin'] >= data['amountMax'])):
-				return Response({"status": 'FAILED'})
+				return Response({"status": 'FAILED', "message":"Les valeurs quantityMin et amountMin ne peuvent pas être supérieur à quantityMax et amountMax"})
 
 			# Récupération de user
 			user = User.objects.get(pseudo=jwt.decode(token, os.environ.get('JWT_SECRET'), algorithms="HS256")['sub'])
@@ -379,9 +417,9 @@ class AdViewset(APIView):
 			if serializer.is_valid():
 				serializer.save()
 				return Response({'status': "SUCCESSFUL"})	
-			return Response({"status": "FAILED"})
+			return Response({'status': "FAILED", "message":"Types des données du JSON invalides!"})
 		except:
-			return Response({"status": "FAILED"})
+			return Response({"status": "FAILED", "message": "Token ou Signature Invalide"})
 		
 	
 	def delete(self, request):
@@ -391,18 +429,18 @@ class AdViewset(APIView):
 		JSON: {
 			'token': '...', 
 			'signature':'...', 
-			'id':'...' // Il s'agit de l'id de l'annonce
+			'id':... // (Il s'agit de l'id de l'annonce) Type Number/int
 		}
 		"""
 		try:
 			data = request.data
 			if len(data) != 3:
-				return Response({"status": "FAILED"})
+				return Response({"status": "FAILED", "message": "JSON invalide"})
 
 			token = data['token'] # Récupération du Token
 			# Vérifie si l'utilisateur est connecté
 			if not isAuthenticated(token, data['signature']):
-				return Response({"status": "FAILED"})
+				return Response({"status": "FAILED", "message":"Vous devez vous connecter"})
 			
 			# Nettoyage de data
 			data.pop('token')
@@ -412,7 +450,7 @@ class AdViewset(APIView):
 			ad.delete()
 			return Response({'status': "SUCCESSFUL"})
 		except:
-			return Response({'status': "FAILED"})
+			return Response({'status': "FAILED", "message":"Token ou Signature invalide"})
 
 
 class AdsViewset(mixins.ListModelMixin, mixins.CreateModelMixin, generics.GenericAPIView):
@@ -423,13 +461,12 @@ class AdsViewset(mixins.ListModelMixin, mixins.CreateModelMixin, generics.Generi
 	def post(self, request, page):
 		try:
 			if page < 1:
-				return Response({"status":"FAILED"})
+				return Response({"status":"FAILED", "message":"L'iindice de page minimal est 1"})
 			token = request.data['token'] # Récupération du Token
 			# Vérifie si l'utilisateur est connecté
 			if not isAuthenticated(token, request.data['signature']):
-				return Response({"status": "FAILED"})
+				return Response({"status": "FAILED", "message":"Vous devez vous connecter"})
 			self.queryset = Ad.objects.order_by('-publicationDate')[ (page-1)*10:10*page]
 			return self.list(request)
 		except:
-			return Response({"status": "FAILED"})
-		
+			return Response({"status": "FAILED", "message":"Token ou Signature Invalide"})
